Refactor: rm queries
